+++ kernel/aster-nix/src/device/tdxguest/mod.rs
@@ -8,7 +8,6 @@ use crate::{
     error::Error,
     events::IoEvents,
     fs::{inode_handle::FileIo, utils::IoctlCmd},
-    prelude::*,
     process::signal::Poller,
 };
